refactor: share MoreOptionsMenu between expanded form modal and panel
